Add more tests for OnDeleteExecute and fix implementation using the result of the ShowYesNoDialog method

+++ FriendStorage/FriendStorage.UI/ViewModel/FriendEditViewModel.cs
@@ -91,9 +91,11 @@ namespace FriendStorage.UI.ViewModel
         private void OnDeleteExecute(object obj)
         {
             var result = _messageDialogService.ShowYesNoDialaog("Delete friend", "");
-
-            _dataProvider.DeleteFriend(Friend.Id);
-            _eventAggregator.GetEvent<FriendDeletedEvent>().Publish(Friend.Id);
+            if (result == MessageDialogResult.Yes)
+            {
+                _dataProvider.DeleteFriend(Friend.Id);
+                _eventAggregator.GetEvent<FriendDeletedEvent>().Publish(Friend.Id);
+            }
         }
 
         private bool OnDeleteCanExecute(object arg)
